requirements.txt
matplotlib==3.2.2
numpy==1.19.5
scikit-learn==0.22.1
torchmetrics==0.10.3
pandas==1.1.5
easydict==1.9
scipy==1.5.3
scikit-image==0.17.2
nibabel==4.0.2
tqdm==4.15.0